requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.9.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.0
pillow==10.2.0
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
sympy==1.12
tokenizers==0.15.2
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.8.0
urllib3==2.2.1
